+++ src/stories/Header.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-// import { Button } from './Button';
 import './header.css';
 
 export const Header = ({ user, onLogin, onLogout, onCreateAccount }) => (
@@ -32,12 +31,10 @@ export const Header = ({ user, onLogin, onLogout, onCreateAccount }) => (
             <span className="welcome">
               Welcome, <b>{user.name}</b>!
             </span>
-            {/* <Button size="small" onClick={onLogout} label="Log out" /> */}
           </>
         ) : (
           <>
-            {/* <Button size="small" onClick={onLogin} label="Log in" /> */}
-            {/* <Button primary size="small" onClick={onCreateAccount} label="Sign up" /> */}
+            Logged out
           </>
         )}
       </div>
